interview-platform/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "alembic>=1.16.5",
    "fastapi>=0.128.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "uvicorn>=0.39.0",
]
